+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added methodology and analysis_intent as reviewable sections in WebUI inline comments
+
 ## [0.2.0] - 2026-03-11
 
 ### Added